pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "automationexercise_ui_selenium_python"
version = "0.1.0"
description = "UI automation framework for AutomationExercise using Selenium + Pytest + POM"
authors = [
    { name = "Your Name", email = "your-email@example.com" }
]
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "selenium>=4.0",
    "pytest>=7.0",
    "allure-pytest",
    "webdriver-manager",
    "PyYAML",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[project.optional-dependencies]
dev = [
    "pytest-xdist",
    "pytest-rerunfailures",
    "flake8",
    "black",
    "isort",
]